Refactor frontend components into three-file pattern (issue #461)
Agent-Logs-Url: https://github.com/darthjee/navi/sessions/d4adf759-39c7-4501-a99f-dc34240dd847
