+++ frontend/src/components/BaseUrlsMenuView.jsx
@@ -1,0 +1,22 @@
+import BaseUrlsClient from '../clients/BaseUrlsClient.js';
+import noop from '../utils/noop.js';
+
+class BaseUrlsMenuView {
+  static buildEffect(setBaseUrls) {
+    return () => {
+      BaseUrlsClient.fetchBaseUrls()
+        .then(setBaseUrls)
+        .catch(noop);
+    };
+  }
+
+  static buildOutsideClickHandler(containerRef, setOpen) {
+    return (e) => {
+      if (containerRef.current && !containerRef.current.contains(e.target)) {
+        setOpen(false);
+      }
+    };
+  }
+}
+
+export default BaseUrlsMenuView;
