+++ frontend/src/components/EngineControls.jsx
@@ -1,23 +1,25 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import EngineControlsHelper from './EngineControlsHelper.jsx';
+import EngineControlsView from './EngineControlsView.jsx';
 
 function EngineControls() {
   const [status, setStatus] = useState(null);
   const intervalRef = useRef(null);
+  const helper = useMemo(() => new EngineControlsHelper(), []);
 
   const refreshStatus = useCallback(
-    () => EngineControlsHelper.fetchStatus(setStatus),
+    () => EngineControlsView.fetchStatus(setStatus),
     []
   );
 
-  const helper = useMemo(
-    () => EngineControlsHelper.build(status, refreshStatus),
+  const view = useMemo(
+    () => EngineControlsView.build(status, refreshStatus),
     [status, refreshStatus]
   );
 
-  useEffect(helper.buildPollingEffect(intervalRef, refreshStatus), [refreshStatus]);
+  useEffect(view.buildPollingEffect(intervalRef, refreshStatus), [refreshStatus]);
 
-  return helper.render();
+  return helper.render(view);
 }
 
 export default EngineControls;
