+++ frontend/src/components/Job.jsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import JobDetails from './JobDetails.jsx';
 import JobHelper from './JobHelper.jsx';
+import JobView from './JobView.jsx';
 
 function Job() {
   const { id } = useParams();
@@ -10,7 +11,7 @@ function Job() {
   const [loading, setLoading] = useState(true);
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  useEffect(JobHelper.buildEffect(id, setJob, setError, setLoading), [id]);
+  useEffect(JobView.buildEffect(id, setJob, setError, setLoading), [id]);
 
   if (loading) return JobHelper.renderLoading();
   if (error) return JobHelper.renderError(error);
