+++ frontend/src/components/JobDetails.jsx
@@ -1,82 +1,7 @@
-import { Link } from 'react-router-dom';
-import CollapsibleSection from './CollapsibleSection.jsx';
-import ReadyCountdown from './ReadyCountdown.jsx';
-import { VARIANT_BY_STATUS } from '../constants/jobStatus.js';
-
-const STATUSES_WITH_REMAINING_ATTEMPTS = new Set(['enqueued', 'processing', 'failed']);
-const STATUSES_WITH_READY_IN = new Set(['failed']);
-const STATUSES_WITH_ERROR = new Set(['failed', 'dead']);
+import JobDetailsHelper from './JobDetailsHelper.jsx';
 
 function JobDetails({ job }) {
-  const variant = VARIANT_BY_STATUS[job.status] ?? 'secondary';
-
-  return (
-    <div className="container mt-4">
-      <h1 className="mb-4">Job Details</h1>
-      <div className="card">
-        <div className="card-body">
-          <dl className="row mb-0">
-            <dt className="col-sm-3">ID</dt>
-            <dd className="col-sm-9 font-monospace">{job.id}</dd>
-
-            <dt className="col-sm-3">Status</dt>
-            <dd className="col-sm-9">
-              <span className={`badge text-bg-${variant}`}>{job.status}</span>
-            </dd>
-
-            <dt className="col-sm-3">Attempts</dt>
-            <dd className="col-sm-9">{job.attempts}</dd>
-
-            <dt className="col-sm-3">Class</dt>
-            <dd className="col-sm-9">{job.jobClass}</dd>
-
-            {job.jobClass === 'ResourceRequestJob' && (
-              <>
-                <dt className="col-sm-3">URL</dt>
-                <dd className="col-sm-9 font-monospace">{job.arguments?.url}</dd>
-              </>
-            )}
-
-            <dt className="col-sm-3">Arguments</dt>
-            <dd className="col-sm-9">
-              <CollapsibleSection label="Show arguments">
-                <pre>{JSON.stringify(job.arguments, null, 2)}</pre>
-              </CollapsibleSection>
-            </dd>
-
-            {STATUSES_WITH_REMAINING_ATTEMPTS.has(job.status) && (
-              <>
-                <dt className="col-sm-3">Remaining attempts</dt>
-                <dd className="col-sm-9">{job.remainingAttempts}</dd>
-              </>
-            )}
-
-            {STATUSES_WITH_READY_IN.has(job.status) && (
-              <>
-                <dt className="col-sm-3">Ready in</dt>
-                <dd className="col-sm-9"><ReadyCountdown readyInMs={job.readyInMs} /></dd>
-              </>
-            )}
-
-            {STATUSES_WITH_ERROR.has(job.status) && job.lastError !== null && job.lastError !== undefined && (
-              <>
-                <dt className="col-sm-3">Last error</dt>
-                <dd className="col-sm-9">
-                  <CollapsibleSection label="Show error">
-                    <p>{job.lastError}</p>
-                    {job.backtrace && <pre>{job.backtrace}</pre>}
-                  </CollapsibleSection>
-                </dd>
-              </>
-            )}
-          </dl>
-        </div>
-      </div>
-      <div className="mt-3">
-        <Link to="/jobs">← Back to Jobs</Link>
-      </div>
-    </div>
-  );
+  return JobDetailsHelper.render(job);
 }
 
 export default JobDetails;
