+++ frontend/src/components/JobHelper.jsx
@@ -1,19 +1,6 @@
 import { Link } from 'react-router-dom';
-import fetchJob from '../clients/JobClient.js';
 
 class JobHelper {
-  static buildEffect(id, setJob, setError, setLoading) {
-    return () => {
-      fetchJob(id)
-        .then((data) => {
-          setJob(data);
-          setError(null);
-        })
-        .catch((err) => setError(err.message))
-        .finally(() => setLoading(false));
-    };
-  }
-
   static renderLoading() {
     return (
       <div className="container mt-5 text-center">
