+++ frontend/src/components/LogsPage.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useMemo, useRef, useState } from 'react';
 import LogsPageHelper from './LogsPageHelper.jsx';
+import LogsPageView from './LogsPageView.jsx';
 import './LogsPage.css';
 
 function LogsPage() {
@@ -9,10 +10,11 @@ function LogsPage() {
   const lastIdRef = useRef(null);
 
   const helper = useMemo(() => LogsPageHelper.build(logs), [logs]);
+  const view = useMemo(() => LogsPageView.build(logs), [logs]);
 
-  useEffect(helper.buildPollingEffect(cancelledRef, lastIdRef, setLogs), []);
+  useEffect(view.buildPollingEffect(cancelledRef, lastIdRef, setLogs), []);
 
-  useEffect(helper.buildScrollEffect(bottomRef), [helper]);
+  useEffect(view.buildScrollEffect(bottomRef), [view]);
 
   return helper.render(bottomRef);
 }
