+++ frontend/src/components/LogsPageHelper.jsx
@@ -1,7 +1,3 @@
-import fetchLogs from '../clients/LogsClient.js';
-
-const POLL_DELAY_MS = 1000;
-
 const LEVEL_CLASS = {
   debug: 'text-debug',
   info: '',
@@ -18,24 +14,6 @@ class LogsPageHelper {
 
   static build(logs) {
     return new LogsPageHelper(logs);
-  }
-
-  buildPollingEffect(cancelledRef, lastIdRef, setLogs) {
-    return () => {
-      cancelledRef.current = false;
-      this.#poll(cancelledRef, lastIdRef, setLogs);
-      return () => {
-        cancelledRef.current = true;
-      };
-    };
-  }
-
-  buildScrollEffect(bottomRef) {
-    return () => {
-      if (this.#logs.length > 0) {
-        bottomRef.current?.scrollIntoView?.({ behavior: 'smooth' });
-      }
-    };
   }
 
   render(bottomRef) {
@@ -60,36 +38,7 @@ class LogsPageHelper {
       </div>
     );
   }
-
-  #hasEntries(entries) {
-    return entries.length !== 0;
-  }
-
-  #appendEntries(entries, lastIdRef, setLogs, cancelledRef) {
-    lastIdRef.current = entries[entries.length - 1].id;
-    setLogs((prev) => [...prev, ...entries]);
-    this.#poll(cancelledRef, lastIdRef, setLogs);
-  }
-
-  #handleEntries(entries, cancelledRef, lastIdRef, setLogs) {
-    if (cancelledRef.current) return;
-
-    if (this.#hasEntries(entries)) {
-      this.#appendEntries(entries, lastIdRef, setLogs, cancelledRef);
-    } else {
-      setTimeout(() => this.#poll(cancelledRef, lastIdRef, setLogs), POLL_DELAY_MS);
-    }
-  }
-
-  #poll(cancelledRef, lastIdRef, setLogs) {
-    if (cancelledRef.current) return;
-
-    fetchLogs({ lastId: lastIdRef.current })
-      .then((entries) => this.#handleEntries(entries, cancelledRef, lastIdRef, setLogs))
-      .catch(() => {
-        if (!cancelledRef.current) setTimeout(() => this.#poll(cancelledRef, lastIdRef, setLogs), POLL_DELAY_MS);
-      });
-  }
 }
 
 export default LogsPageHelper;
+
